build: use rust 1.82 and fix clipyy errors

+++ docx-core/src/documents/elements/line_spacing.rs
@@ -52,7 +52,7 @@ impl LineSpacing {
     }
 
     pub fn line(mut self, line: i32) -> Self {
-        self.line = Some(line as i32);
+        self.line = Some(line);
         self
     }
 }
